Make jumpstart helper and unignore other test
