+++ crates/threshold-signature-server/src/helpers/tests.rs
@@ -39,14 +39,17 @@ use crate::{
     AppState,
 };
 use axum::{routing::IntoMakeService, Router};
+use entropy_client::substrate::query_chain;
 use entropy_kvdb::{encrypted_sled::PasswordMethod, get_db_path, kv_manager::KvManager};
 use entropy_protocol::PartyId;
 #[cfg(test)]
 use entropy_shared::EncodedVerifyingKey;
-use entropy_shared::{EVE_VERIFYING_KEY, NETWORK_PARENT_KEY};
+use entropy_shared::{OcwMessageDkg, EVE_VERIFYING_KEY, NETWORK_PARENT_KEY};
+use futures::future::join_all;
+use parity_scale_codec::Encode;
 use std::time::Duration;
 use subxt::{
-    backend::legacy::LegacyRpcMethods, ext::sp_core::sr25519, tx::PairSigner,
+    backend::legacy::LegacyRpcMethods, events::EventsClient, ext::sp_core::sr25519, tx::PairSigner,
     utils::AccountId32 as SubxtAccountId32, Config, OnlineClient,
 };
 use tokio::sync::OnceCell;
@@ -342,4 +345,63 @@ pub async fn store_program_and_register(
     .unwrap();
 
     (verifying_key, program_hash)
+}
+
+pub async fn put_jumpstart_request_on_chain(
+    api: &OnlineClient<EntropyConfig>,
+    rpc: &LegacyRpcMethods<EntropyConfig>,
+    pair: sr25519::Pair,
+) {
+    let account = PairSigner::<EntropyConfig, sp_core::sr25519::Pair>::new(pair);
+
+    let registering_tx = entropy::tx().registry().jump_start_network();
+    submit_transaction(api, rpc, &account, &registering_tx, None).await.unwrap();
+}
+
+pub async fn do_jump_start(
+    api: &OnlineClient<EntropyConfig>,
+    rpc: &LegacyRpcMethods<EntropyConfig>,
+    pair: sr25519::Pair,
+) {
+    let block_number = rpc.chain_get_header(None).await.unwrap().unwrap().number + 1;
+    put_jumpstart_request_on_chain(&api, &rpc, pair).await;
+
+    run_to_block(&rpc, block_number + 1).await;
+
+    let selected_validators_query = entropy::storage().registry().jumpstart_dkg(block_number);
+    let validators_info =
+        query_chain(&api, &rpc, selected_validators_query, None).await.unwrap().unwrap();
+    let validators_info: Vec<_> = validators_info.into_iter().map(|v| v.0).collect();
+    let onchain_user_request = OcwMessageDkg { block_number, validators_info };
+
+    let client = reqwest::Client::new();
+    let response_results = join_all(
+        vec![3002, 3003, 3004]
+            .iter()
+            .map(|port| {
+                client
+                    .post(format!("http://127.0.0.1:{}/generate_network_key", port))
+                    .body(onchain_user_request.clone().encode())
+                    .send()
+            })
+            .collect::<Vec<_>>(),
+    )
+    .await;
+    for response_result in response_results {
+        assert_eq!(response_result.unwrap().text().await.unwrap(), "");
+    }
+
+    // Wait for jump start event
+    let mut got_jumpstart_event = false;
+    for _ in 0..75 {
+        std::thread::sleep(std::time::Duration::from_millis(1000));
+        let block_hash = rpc.chain_get_block_hash(None).await.unwrap();
+        let events = EventsClient::new(api.clone()).at(block_hash.unwrap()).await.unwrap();
+        let jump_start_event = events.find::<entropy::registry::events::FinishedNetworkJumpStart>();
+        for _event in jump_start_event.flatten() {
+            got_jumpstart_event = true;
+            break;
+        }
+    }
+    assert!(got_jumpstart_event);
 }
